mapper/requirements.txt
functions-framework==3.*
google-cloud-storage==2.13.0
six==1.16.0